Data type changed

+++ client/math/dataTransmission.go
@@ -9,7 +9,7 @@ import (
 )
 
 // Transmit sends measurements results to server
-func Transmit(client *http.Client, array []uint32, time_array []float64) (*http.Response, error) {
+func Transmit(client *http.Client, array []float64, time_array []float64) (*http.Response, error) {
 	dataMsg := protocol.MathData{
 		DataArray: array,
 		TimeArray: time_array}
